pyproject.toml
[project]
name = "productivity-app"
version = "0.1.0"
description = "Pomodoro timer + todo list with Obsidian daily note sync"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
